Refactor file handling by removing fileDrawerStore and integrating FileSidebar directly into ExampleComponent. Updated layout to remove unused FileSidebar import. Enhanced FileSidebar to manage file selection and state locally.
